월요일 데모 후 추가사항들...
client
	-스타일 수정
	-메인 페이지 글자 추가
	-게시글 삭제시 confirm 추가
	-회원가입시 userService.join 실행 (빠뜨렸었음..)

server
	- contents model 추가
	- contents 삭제/수정시 유저 아이디와 jwt 로 확인된 유저 아이디 비교하기
	- get top5 추가 ( /src/config/summarizer.ts)

+++ client/src/app/contents-posting/contents-posting.component.ts
@@ -60,8 +60,10 @@ export class ContentsPostingComponent implements OnInit {
 
 
   deleteContent():void{
-    this.contentService.deleteContent(this.id);
-    this.router.navigate(['/contents-list']); //redirect
+    if(window.confirm('삭제하시겠습니까?') ){
+      this.contentService.deleteContent(this.id);
+      this.router.navigate(['/contents-list']); //redirect
+    }
   }
 
   onSubmit():void{
